requirements.txt
anthropic>=0.40.0
openai>=1.30.0
keyboard>=0.13.5
mss>=9.0.1
Pillow>=10.0.0
pystray>=0.19.5
python-dotenv>=1.0.0
